requirements.txt
streamlit
pytesseract
Pillow
langchain
langchain_groq
dotenv
fpdf